pyproject.toml
[tool.poetry]
name = "the-flask-test"
version = "0.1.0"
description = ""
authors = ["shekhar-shubhendu"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1"
flask_sqlalchemy = "^2.4"
flask-marshmallow = "^0.10.1"
marshmallow-sqlalchemy = "^0.21.0"
Flask-Migrate = "^2.5"
python-dotenv = "^0.10.5"
gunicorn = "^20.0.4"
psycopg2 = "^2.8.4"
flask-cors = "^3.0.8"

[tool.poetry.dev-dependencies]
pytest = "^5.3"
flake8 = "^3.7"

[tool.dephell.main]
from = {format = "poetry", path = "pyproject.toml"}
to = {format = "pip", path = "requirements.txt"}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
